frontend/requirements.txt
flask
werkzeug
python-dotenv
elasticsearch